% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{describeArrowArray}
\alias{describeArrowArray}
\title{Describe An Arrow Array}
\usage{
describeArrowArray(ptr)
}
\arguments{
\item{ptr}{An external pointer object for an \code{ArrowArray} structure}
}
\value{
None. The function is invoked for the side effect of displaingy some
array info.
}
\description{
Describe An Arrow Array
}
