% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{extract_null_count}
\alias{extract_null_count}
\title{Count Null Elements in Arrow Object}
\usage{
extract_null_count(array_data_xptr)
}
\arguments{
\item{array_data_xptr}{An external pointer to an Arrow object}
}
\value{
The number of null elements
}
\description{
Count Null Elements in Arrow Object
}
\examples{
extract_null_count(narrow::as_narrow_array(c(NA, NA, 1:5))$array_data)
}
