% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{narrow_array_info}
\alias{narrow_array_info}
\title{Call array info}
\usage{
narrow_array_info(adptr)
}
\arguments{
\item{adxpt}{An external pointer to array data object}
}
\value{
A list with array info
}
\description{
Call array info
}
\examples{
x <- narrow::as_narrow_array(c(NA, 2, NA, 4:7))
narrow_array_info(x$array_data)
}
