% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{print_uint64}
\alias{print_uint64}
\title{Print uint64_t vector}
\usage{
print_uint64(ptr)
}
\arguments{
\item{ptr}{An external pointer object for an \code{uint64_t} vector}
}
\value{
None. The function is invoked for the side effect of display schema info.
}
\description{
Print uint64_t vector
}
