\name{RcppArrow-package}
\alias{RcppArrow-package}
\alias{RcppArrow}
\docType{package}
\title{\packageTitle{RcppArrow}}
\description{\packageDescription{RcppArrow}}
\section{Package Content}{\packageIndices{RcppArrow}}
\author{\packageAuthor{RcppArrow}}
\section{Maintainer}{\packageMaintainer{RcppArrow}}
\keyword{package}
