\name{RcppNarrow-package}
\alias{RcppNarrow-package}
\alias{RcppNarrow}
\docType{package}
\title{\packageTitle{RcppNarrow}}
\description{\packageDescription{RcppNarrow}}
\section{Package Content}{\packageIndices{RcppNarrow}}
\author{\packageAuthor{RcppNarrow}}
\section{Maintainer}{\packageMaintainer{RcppNarrow}}
\keyword{package}
