% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{describeArrowSchema}
\alias{describeArrowSchema}
\title{Describe An Arrow Schema}
\usage{
describeArrowSchema(ptr)
}
\arguments{
\item{ptr}{An external pointer object for an \code{ArrowSchema} structure}
}
\value{
None. The function is invoked for the side effect of displaying the
schema info.
}
\description{
Describe An Arrow Schema
}
